requirements.txt
opentelemetry-exporter-jaeger-thrift==1.1.0
opentelemetry-instrumentation-asgi==0.20b
opentelemetry-instrumentation-fastapi==0.20b
fastapi==0.63.0
uvicorn[standard]==0.13.4
